xtred/requirements.txt
Django>=4.2.15
python-dotenv==1.0.1
django-bootstrap5==22.2
django_debug_toolbar==3.8.1
djangorestframework==3.15.2
djoser==2.1.0
django-filter==23.1
drf-yasg==1.21.7
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
pillow==10.3.0
flake8-isort==6.0.0
black==24.8.0
isort==5.13.2
mypy==1.11.2